backend/requirements.txt
Django==1.11.7
django-extensions==1.9.7
dj-database-url==0.4.2
djangorestframework==3.7.1
mysqlclient==1.3.12
httpie==0.9.9
